gh-409, gh-409 gh-414: Dockerize JanusGraph
This commits adds a dockerized JanusGraph. Closes #409.

We add configuration to JanusGraph so that it will use
an embedded Cassandra database (in the same JVM), and use the Lucene
library for search in the same JVM too.

This makes the startup of all components easier than when these
functionalities are provided by two different programs.

We need to have a full configuration (external Cassandra
cluster and an Elastic Search cluster) later on, but for the inital
development of the JanusGraph connector, this will suffice.